@@ -5,13 +5,15 @@ import (
 	"encoding/json"
 	"flag"
 	"fmt"
-	"io/ioutil"
 	"os"
+	"strings"
 
 	"github.com/fatih/color"
 	"github.com/grafana/plugin-validator/pkg/analysis"
 	"github.com/grafana/plugin-validator/pkg/analysis/passes"
 	"github.com/grafana/plugin-validator/pkg/archivetool"
+	"github.com/grafana/plugin-validator/pkg/logme"
+	"github.com/grafana/plugin-validator/pkg/repotool"
 	"github.com/grafana/plugin-validator/pkg/runner"
 	"gopkg.in/yaml.v2"
 )
@@ -24,20 +26,27 @@ type FormattedOutput struct {
 
 func main() {
 	var (
-		strictFlag = flag.Bool("strict", false, "If set, plugincheck returns non-zero exit code for warnings")
-		configFlag = flag.String("config", "", "Path to configuration file")
+		strictFlag    = flag.Bool("strict", false, "If set, plugincheck returns non-zero exit code for warnings")
+		configFlag    = flag.String("config", "", "Path to configuration file")
+		sourceCodeUri = flag.String("sourceCodeUri", "", "URL to the source code of the plugin. If set, the source code will be downloaded and analyzed. This can be a ZIP file or an URL to git repository")
 	)
 
 	flag.Parse()
 
+	logme.Debugln("stric mode: ", *strictFlag)
+	logme.Debugln("config file: ", *configFlag)
+	logme.Debugln("source code: ", *sourceCodeUri)
+	logme.Debugln("archive file: ", flag.Arg(0))
+
 	if *configFlag == "" {
-		fmt.Fprintln(os.Stderr, "missing config")
+		logme.Errorln("no config file specified")
+		flag.Usage()
 		os.Exit(1)
 	}
 
 	cfg, err := readConfigFile(*configFlag)
 	if err != nil {
-		fmt.Fprintln(os.Stderr, fmt.Errorf("couldn't read configuration: %w", err))
+		logme.Errorln(fmt.Errorf("couldn't read configuration: %w", err))
 		os.Exit(1)
 	}
 
@@ -50,21 +59,30 @@ func main() {
 
 	b, err := archivetool.ReadArchive(pluginURL)
 	if err != nil {
-		fmt.Fprintln(os.Stderr, fmt.Errorf("couldn't fetch plugin archive: %w", err))
+		logme.Errorln(fmt.Errorf("couldn't fetch plugin archive: %w", err))
 		os.Exit(1)
 	}
 
 	// Extract the ZIP archive in a temporary directory.
-	archiveDir, cleanup, err := archivetool.ExtractPlugin(bytes.NewReader(b))
+	archiveDir, archiveCleanup, err := archivetool.ExtractPlugin(bytes.NewReader(b))
 	if err != nil {
-		fmt.Fprintln(os.Stderr, fmt.Errorf("couldn't extract plugin archive: %w", err))
+		logme.Errorln(fmt.Errorf("couldn't extract plugin archive: %w", err))
 		os.Exit(1)
 	}
-	defer cleanup()
+	defer archiveCleanup()
 
-	diags, err := runner.Check(passes.Analyzers, archiveDir, cfg)
+	sourceCodeDir, sourceCodeDirCleanup, err := getSourceCodeDir(*sourceCodeUri)
 	if err != nil {
-		fmt.Fprintln(os.Stderr, fmt.Errorf("check failed: %w", err))
+		// if source code is not provided, we don't fail the validation
+		logme.Errorln(fmt.Errorf("couldn't get source code: %w", err))
+	}
+	if sourceCodeDirCleanup != nil {
+		defer sourceCodeDirCleanup()
+	}
+
+	diags, err := runner.Check(passes.Analyzers, archiveDir, sourceCodeDir, cfg)
+	if err != nil {
+		logme.Errorln(fmt.Errorf("check failed: %w", err))
 		os.Exit(1)
 	}
 
@@ -133,11 +151,12 @@ func main() {
 		}
 	}
 
+	logme.DebugFln("exit code: %d", exitCode)
 	os.Exit(exitCode)
 }
 
 func readConfigFile(path string) (runner.Config, error) {
-	b, err := ioutil.ReadFile(path)
+	b, err := os.ReadFile(path)
 	if err != nil {
 		return runner.Config{}, err
 	}
@@ -148,4 +167,37 @@ func readConfigFile(path string) (runner.Config, error) {
 	}
 
 	return config, nil
+}
+
+func getSourceCodeDir(sourceCodeUri string) (string, func(), error) {
+	// If source code URI is not provided, return immediately with an empty string
+	// otherwise we will get an error when trying to extract the source code archive
+	if sourceCodeUri == "" {
+		return "", func() {}, nil
+	}
+
+	// file:// protocol for local directories
+	if strings.HasPrefix(sourceCodeUri, "file://") {
+		sourceCodeDir := strings.TrimPrefix(sourceCodeUri, "file://")
+		if _, err := os.Stat(sourceCodeDir); err != nil {
+			return "", nil, err
+		}
+		return sourceCodeDir, func() {}, nil
+	}
+
+	if repotool.IsSupportedGitUrl(sourceCodeUri) {
+		extractedGitRepo, sourceCodeCleanUp, err := repotool.GitUrlToLocalPath(sourceCodeUri)
+		if err != nil {
+			return "", sourceCodeCleanUp, err
+		}
+		return extractedGitRepo, sourceCodeCleanUp, nil
+	}
+
+	// assume is an archive url
+	extractedDir, sourceCodeCleanUp, err := archivetool.ArchiveToLocalPath(sourceCodeUri)
+	if err != nil {
+		return "", sourceCodeCleanUp, fmt.Errorf("couldn't extract source code archive: %s. %w", sourceCodeUri, err)
+	}
+	return extractedDir, sourceCodeCleanUp, nil
+
 }
