@@ -56,8 +56,8 @@ func TestNoLicenseFound(t *testing.T) {
 	_, err := Analyzer.Run(pass)
 	require.NoError(t, err)
 	require.Len(t, interceptor.Diagnostics, 1)
-	require.Equal(t, interceptor.Diagnostics[0].Title, "License not found")
-	require.Equal(t, interceptor.Diagnostics[0].Detail, "Could not find or parse the license file inside the plugin archive. Please make sure to include a LICENCE file in your archive.")
+	require.Equal(t, interceptor.Diagnostics[0].Title, "LICENSE file not found")
+	require.Equal(t, interceptor.Diagnostics[0].Detail, "Could not find a license file inside the plugin archive. Please make sure to include a LICENCE file in your archive.")
 }
 
 func TestValidMitLicense(t *testing.T) {
@@ -91,5 +91,5 @@ func TestInvalidICsLicense(t *testing.T) {
 	require.NoError(t, err)
 	require.Len(t, interceptor.Diagnostics, 1)
 	require.Equal(t, interceptor.Diagnostics[0].Title, "Valid license not found")
-	require.Equal(t, interceptor.Diagnostics[0].Detail, "Could not find a license file inside the plugin archive or the provided license is not compatible with Grafana plugins. Please refer to https://grafana.com/licensing/ for more information.")
+	require.Equal(t, interceptor.Diagnostics[0].Detail, "The provided license is not compatible with Grafana plugins. Please refer to https://grafana.com/licensing/ for more information.")
 }
